finished url-getters for iconhashes

+++ include/dpp/guild.h
@@ -223,7 +223,7 @@ public:
 	 * @brief Returns the members's per guild avatar if they have one, otherwise returns an empty string
 	 * 
 	 * @param size The size of the avatar in pixels. It can be any power of two between 16 and 4096. If not specified, the default sized avatar is returned.
-	 * @return std::string avatar or empty string
+	 * @return std::string avatar url or empty string
 	 */
 	std::string get_avatar_url(uint16_t size = 0) const;
 
@@ -546,37 +546,37 @@ public:
 	 */
 	bool connect_member_voice(snowflake user_id, bool self_mute = false, bool self_deaf = false);
 
-	/**
-	 * @brief Get the banner url of the guild object
+    /**
+	 * @brief Get the banner url of the guild if it have one, otherwise returns an empty string
 	 *
-	 * @param size The size of the banner in pixels, if not specified, the default size is returned.
-	 * @return std::string banner url
+	 * @param size The size of the banner in pixels. It can be any power of two between 16 and 4096. if not specified, the default sized banner is returned.
+	 * @return std::string banner url or empty string
 	 */
-	std::string get_banner_url(uint32_t size = 0) const;
+    std::string get_banner_url(uint16_t size = 0) const;
 
-	/**
-	 * @brief Get the discovery splash url of the guild object
+    /**
+	 * @brief Get the discovery splash url of the guild if it have one, otherwise returns an empty string
 	 *
-	 * @param size The size of the discovery splash in pixels, if not specified, the default size is returned.
-	 * @return std::string discovery splash url
+	 * @param size The size of the discovery splash in pixels. It can be any power of two between 16 and 4096. if not specified, the default sized discovery splash is returned.
+	 * @return std::string discovery splash url or empty string
 	 */
-	std::string get_discovery_splash_url(uint32_t size = 0) const;
+    std::string get_discovery_splash_url(uint16_t size = 0) const;
 
-	/**
-	 * @brief Get the icon url of the guild object
+    /**
+	 * @brief Get the icon url of the guild if it have one, otherwise returns an empty string
 	 *
-	 * @param size The size of the icon in pixels, if not specified, the default size is returned.
-	 * @return std::string icon url
+	 * @param size The size of the icon in pixels. It can be any power of two between 16 and 4096. if not specified, the default sized icon is returned.
+	 * @return std::string icon url or empty string
 	 */
-	std::string get_icon_url(uint32_t size = 0) const;
+    std::string get_icon_url(uint16_t size = 0) const;
 
-	/**
-	 * @brief Get the splash url of the guild object
+    /**
+	 * @brief Get the splash url of the guild if it have one, otherwise returns an empty string
 	 *
-	 * @param size The size of the splash in pixels, if not specified, the default size is returned.
-	 * @return std::string splash url
+	 * @param size The size of the splash in pixels. It can be any power of two between 16 and 4096. if not specified, the default sized splash is returned.
+	 * @return std::string splash url or empty string
 	 */
-	std::string get_splash_url(uint32_t size = 0) const;
+    std::string get_splash_url(uint16_t size = 0) const;
 
 	/**
 	 * @brief Set the name of the guild in the object
